updated scripts and created summary notebook
