+++ README.md
@@ -3,20 +3,32 @@
 ## About
 This repository was created as part of an assignment in [UBC's Master of Data Science](https://masterdatascience.ubc.ca/) program.
 
+## Usage
+Load the data:
+`python3 scripts/load_data.py --url="https://github.com/ageron/handson-ml/blob/master/datasets/housing/housing.csv?raw=true" --file_path="data/raw_data.csv"`
+
+Complete Wrangling and Split into Training and Testing sets:
+`Rscript scripts/wrangle-and-split-data.R --filepath_in='data/raw_data.csv' --filepath_out_train='data/train.csv' --filepath_out_test='data/test.csv'`
+
+Download figures as part of an analysis completed:
+`python3 scripts/eda_v2.py --train_path='data/train.csv' --out_folder_path='results/eda_charts/'`
+
+Machine Learning Results:
+`python3 scripts/ML_analysis_v2.py --training_input_path='data/train.csv' --testing_input_path='data/test.csv'  --output_path='results/ml_results'`
+
+The final report can be found [here](results/california_housing_predict_report.ipynb).
+
 ## Proposal
 
 #### Dataset
 This dataset is a modified version of [The California Housing Dataset](https://www.dcc.fc.up.pt/~ltorgo/Regression/cal_housing.html), with [additional columns added by Aurélien Geron](https://github.com/ageron/handson-ml). This dataset contains information about median California house values (including related factors) as sourced from the 1990 US Census.
-
-The script to load the California Housing Dataset into python can be found in [scripts](/scripts).
 
 #### Research Question
 As discussed in a special report by The Economist on January 16, 2020 ['Housing is at the root of many of the rich world’s problems'](https://www.economist.com/special-report/2020/01/16/housing-is-at-the-root-of-many-of-the-rich-worlds-problems) and nowhere is this problem more acute than in the richest and most populous American state where Bloomberg has attempted to explain ['How California Became America’s Housing Market Nightmare'](https://www.bloomberg.com/graphics/2019-california-housing-crisis/).
 
 Although finding solutions for this deep-seated socioeconomic problem is beyond the scope of this research project, we propose to build a machine learning model using this dataset to identify factors which can be used to predict housing prices and help to better inform stakeholders in this increasingly unaffordable market. Based on the latest report [California Housing Affordability Update - Q3 - 2019](https://www.car.org/marketdata/data/haitraditional) from the California Association of Realtors, only 31% of California residents can afford to purchase the median price home in their region.
 
-Our research will attempt to identify which machine learning model and model coefficients are the best predictors of median house prices in California based on the 10 variables from the dataset shown below which contains 20,640 examples. Our goal is to build a model with significantly higher accuracy than the 0.60 achieved by [Eric Chen](https://www.kaggle.com/ericfeng84), the author of [The California House Price](https://www.kaggle.com/ericfeng84/the-california-housing-price) Kaggle page from which the dataset was obtained.
-
+Our research focus is to predict the median housing price per census block in California based on the nine given variables:
 ##### Location
 * longitude: longitude of the census block
 * latitude: latitude of the census block
@@ -34,16 +46,15 @@ Our research will attempt to identify which machine learning model and model coe
 
 ##### Target
 * median_house_value
-  
-  
+
+Our goal is to build a model that will predict median house value with a higher model score than the 0.60 achieved by [Eric Chen](https://www.kaggle.com/ericfeng84), the author of [The California House Price](https://www.kaggle.com/ericfeng84/the-california-housing-price) Kaggle page from which the dataset was obtained.
+
 #### How we will analyze the Data
 Since this dataset is publicly available on [Kaggle](https://www.kaggle.com/camnugent/california-housing-prices/kernels), there have been many analyses completed. The analysis that we focused on was [completed by Aurélien Geron](https://www.kaggle.com/artemiosgeromitsos/housing-super-duper-kernel/notebook). He concluded that a linear regressor was the most appropriate model. The focus of our analysis is to attempt to improve on his success by taking into consideration multicollinearity (relationships that exist between independent variables), tuning hyperparameters (hyperparameters will be chosen based on GridSearchCV, which uses the given estimator's default scoring method), and trying different models that were not included in his analysis.
 
-We will use Linear Regression, K-Nearest Neighbour Trees, Random Forests, and Support Vector Regressors to predict the median house value given the independent variables.
+We will use Linear Regression, K-Nearest Neighbour Trees, and a Random Forest Regressor to predict the median house value given the independent variables.
 
 #### Exploratory Data Analysis Figures and Tables to be Used
-The full Exploratory Data Analysis can be found [here](analysis/EDA.ipynb).
-
 Before diving into the figures, it is key to understand that each record (row) of data represents a [census block](https://www.census.gov/newsroom/blogs/random-samplings/2011/07/what-are-census-blocks.html) in California, and some attributes describe a census block (like the population of the given census block), while others describe medians within a census block (median income in the census block).
 
 The first set of figures produced shows each of the quantitative explanatory variables plotted against the median house value. These 6 plots were chosen as an initial analysis because in Aurélien Geron's version of this analysis, he concluded that linear regression was the best estimation method. We chose to start by attempting to see which (if any) of the explanatory variables had patterns or variances compared to median house price.
