+++ results/eda_charts/vif_table.csv
@@ -1,0 +1,8 @@
+variable,vif_val
+housing_median_age,1.1639048179556963
+total_rooms,11.84944300462276
+total_bedrooms,34.89104735218172
+population,6.5828368994484405
+households,33.871692634309554
+median_income,1.5242627814774108
+intercept,18.278944292285182
