+++ results/ml_results/knn_results_table.csv
@@ -1,0 +1,20 @@
+n_neighbours,train_error,test_error
+1,0.0,0.417
+2,0.105,0.324
+3,0.147,0.293
+4,0.169,0.282
+5,0.187,0.271
+6,0.196,0.264
+7,0.205,0.261
+8,0.212,0.26
+9,0.218,0.26
+10,0.222,0.258
+11,0.226,0.258
+12,0.23,0.258
+13,0.233,0.258
+14,0.236,0.258
+15,0.239,0.258
+16,0.241,0.259
+17,0.244,0.26
+18,0.246,0.26
+19,0.248,0.26
