+++ results/ml_results/lr_rfe_results_table.csv
@@ -1,0 +1,10 @@
+n_features_to_select,train_error,test_error
+1,0.978,0.982
+2,0.757,0.76
+3,0.416,0.414
+4,0.413,0.412
+5,0.376,0.394
+6,0.375,0.395
+7,0.372,0.391
+8,0.36,0.376
+9,0.36,0.376
