+++ scripts/load_data.py
@@ -3,7 +3,9 @@
 
 """
 Downloads data csv data from the web to a local filepath as either a csv or feather file format.
-Usage: load_data.py --url=<url> --file_path=<file_path> 
+
+Usage: scripts/load_data.py --url=<url> --file_path=<file_path> 
+
 Options:
 --url=<file_path>  url from where to download the data
 --file_path=<file_path>  Path (including filename) to the csv file.
